backend/pyproject.toml
[project]
name = "wdms-backend"
version = "0.1.0"
description = "Water Delivery Management System — Backend API (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Settings
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "email-validator>=2.2.0",

    # Database
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.2.0",

    # Redis / cache / rate limit
    "redis>=5.0.8",
    "slowapi>=0.1.9",

    # Background jobs
    "celery>=5.4.0",

    # Security
    "argon2-cffi>=23.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.9.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "python-dateutil>=2.9.0",
    "orjson>=3.10.0",

    # Observability
    "sentry-sdk[fastapi]>=2.13.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-python-dateutil",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4", "SIM", "RUF"]
ignore = ["S101", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
